Added support for re-usage and added common assertion point with Util.
